Big warning with instruction on how to start the external ML worker
